requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
pyyaml==6.0.1
apscheduler==3.10.4
pytz==2024.1

